app/requirements.txt
Django==5.0.1
djangorestframework==3.14, <3.15
django-nested-admin==4.0.2
python-dotenv==1.0.0
ruff==0.1.13
Pillow==10.0.0
django-extensions==3.2.3
ipython==8.20.0
django-debug-toolbar==4.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.6
whitenoise==6.5.0
requests==2.31.0
celery[redis]
redis
PyMovieDb==0.0.9
